# cached ShardIds generated by HashCodeMessageExtractor