--- a/src/contrib/cluster/Akka.Cluster.Sharding/ClusterSharding.cs
+++ b/src/contrib/cluster/Akka.Cluster.Sharding/ClusterSharding.cs
@@ -60,6 +60,8 @@ namespace Akka.Cluster.Sharding
         /// </summary>
         public readonly int MaxNumberOfShards;
 
+        private readonly Dictionary<int, string> _cachedIds;
+
         /// <summary>
         /// TBD
         /// </summary>
@@ -67,6 +69,11 @@ namespace Akka.Cluster.Sharding
         protected HashCodeMessageExtractor(int maxNumberOfShards)
         {
             MaxNumberOfShards = maxNumberOfShards;
+            _cachedIds = new Dictionary<int, string>(MaxNumberOfShards);
+            foreach (var c in Enumerable.Range(0, maxNumberOfShards))
+            {
+                _cachedIds[c] = c.ToString();
+            }
         }
 
         /// <summary>
@@ -98,8 +105,8 @@ namespace Akka.Cluster.Sharding
                 id = se.EntityId;
             else
                 id = EntityId(message);
-
-            return (Math.Abs(MurmurHash.StringHash(id)) % MaxNumberOfShards).ToString();
+            
+            return _cachedIds[(Math.Abs(MurmurHash.StringHash(id)) % MaxNumberOfShards)];
         }
     }