@Aaronontheweb 2021-08-19
Doesn't make a huge impact in the aggregate benchmarks, however:

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1165 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  Job-MBAVSR : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT

InvocationCount=1  UnrollFactor=1  

```
|                              Method |   StateMode | MsgCount |         Mean |      Error |     StdDev |
|------------------------------------ |------------ |--------- |-------------:|-----------:|-----------:|
|  **SingleRequestResponseToLocalEntity** | **Persistence** |    **10000** |   **115.776 ms** |  **2.2927 ms** |  **4.6834 ms** |
|              StreamingToLocalEntity | Persistence |    10000 |     4.884 ms |  0.2643 ms |  0.7191 ms |
| SingleRequestResponseToRemoteEntity | Persistence |    10000 | 4,199.020 ms | 30.7013 ms | 27.2159 ms |
|             StreamingToRemoteEntity | Persistence |    10000 |   505.833 ms | 10.0748 ms | 14.7676 ms |
|  **SingleRequestResponseToLocalEntity** |       **DData** |    **10000** |   **113.480 ms** |  **2.2588 ms** |  **4.7645 ms** |
|              StreamingToLocalEntity |       DData |    10000 |     4.785 ms |  0.3805 ms |  1.0918 ms |
| SingleRequestResponseToRemoteEntity |       DData |    10000 | 4,189.848 ms | 78.5903 ms | 80.7064 ms |
|             StreamingToRemoteEntity |       DData |    10000 |   499.897 ms |  8.4260 ms |  7.4695 ms |